pyproject.toml
[project]
name = "crawl-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "crawl4ai>=0.7.8",
    "cryptography>=46.0.3",
    "fastapi>=0.124.4",
    "httpx>=0.27.0",
    "google>=3.0.0",
    "google-genai>=1.59.0",
    "PyJWT>=2.8.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "llama-index>=0.14.12",
    "numpy>=2.3.5",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "pillow>=12.0.0",
    "pycryptodome>=3.23.0",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "rq>=2.6.1",
    "tortoise-orm[asyncpg]>=0.25.1",
    "uvicorn>=0.38.0",
    "yarl>=1.22.0",
    "langchain-core>=1.2.5",
    "shopifyapi>=12.7.0",
]
